v1.0

+++ src/uninstall.sh
@@ -16,23 +16,30 @@ gray="\e[90m"
 clear
 
 echo ""
-echo -e "  ${blue}Keift ${green}Zapret Uninstaller${reset}"
+echo -e "  ${blue}Keift ${cyan}Zapret Uninstaller${reset}"
 echo ""
 
+if ! command -v systemd &>/dev/null; then
+  echo -e "  ${red}Error: It only works on devices where Systemd is installed.${reset}"
+  echo ""
+
+  exit 1
+fi
+
 if [ ! -d "/opt/zapret" ]; then
-  echo -e "  ${gray}Zapret already not installed."
+  echo -e "  ${gray}Zapret already not installed.${reset}"
   echo ""
 
   exit 0
 fi
 
-echo -e "  ${gray}Uninstalling Zapret..."
+echo -e "  ${gray}Uninstalling Zapret...${reset}"
 
 printf "\n" | sudo /opt/zapret/uninstall_easy.sh &>/dev/null
 
 sudo rm -rf /opt/zapret
 sudo rm -rf /tmp/zapret-v72.7
 
-echo -e "  ${gray}Zapret has been successfully uninstalled."
+echo -e "  ${gray}Zapret has been successfully uninstalled.${reset}"
 
 echo ""
